Create data store to handle with Pokémon data
This is a big refactory. This change was very difficult to make without breaking behaviors, which is why this commit is so large.

The state of Pokémon, teams, and targets is now centralized in a data store, making all changes reactive and reducing the complexity of communication between components.

The storage is not working yet and some things still need to be refactored.