Fix error on Analytic ability calc
The behavior was wrong. It was causing an extra 30% increase due to the logic of the function that increases the status of paradoxes. When the logic of paradoxes was adjusted to only affect them, this calculation broke. This ability should not be activated manually, but rather based on the speed of the Pokémon involved.